+++ src/services/recommends/type.ts
@@ -1,0 +1,9 @@
+import { RowDataPacket } from "mysql2";
+
+export interface Recommend extends RowDataPacket {
+  id: number;
+  page: string;
+  type: string;
+  value: string;
+  label: string;
+}
